+++ docs/changelog.rst
@@ -31,6 +31,7 @@ Additions to the :doc:`supporting modules <supporting-modules>`.
     * :py:func:`khoros.utils.tests.test_users.set_package_path`
     * :py:func:`khoros.utils.tests.test_users.test_impersonate_user`
     * :py:func:`khoros.utils.tests.test_users.test_create_user`
+    * :py:func:`khoros.utils.tests.test_users.test_failed_create_user`
 * Added the following functions to the :py:mod:`khoros.utils.tests.test_tags` module:
     * :py:func:`khoros.utils.tests.test_tags.test_invalid_payload_for_single_tag`
     * :py:func:`khoros.utils.tests.test_tags.test_add_single_tag_to_message`